# doc/openapi/swagger.yaml
basePath: /
definitions:
...
  model.CreateUserRequest:
    properties:
      ID:
        type: string
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.UserContent'
      metadata:
        $ref: '#/definitions/model.UserMetadata'
    type: object
...
  model.ListUsersContentResponse:
    properties:
      userContentResponses:
        items:
          $ref: '#/definitions/model.UserContentResponse'
        type: array
    type: object
...
  model.PROJECT_ROLE_TYPES:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    type: integer
    x-enum-varnames:
    - PROJECT_ROLE_TYPES_UNSPECIFIED
    - PROJECT_ROLE_TYPES_PM
    - PROJECT_ROLE_TYPES_EM
    - PROJECT_ROLE_TYPES_DEV
    - PROJECT_ROLE_TYPES_QA
    - PROJECT_ROLE_TYPES_BA
    - PROJECT_ROLE_TYPES_UX
    - PROJECT_ROLE_TYPES_O
...
  model.UpdateUserContentRequest:
    properties:
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.UserContent'
    type: object
  model.UpdateUserMetadataRequest:
    properties:
      UUID:
        type: string
      metadata:
        $ref: '#/definitions/model.UserMetadata'
    type: object
  model.UpdateUserRequest:
    properties:
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.UserContent'
      metadata:
        $ref: '#/definitions/model.UserMetadata'
    type: object
  model.User:
    properties:
      ID:
        type: string
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.UserContent'
      metadata:
        $ref: '#/definitions/model.UserMetadata'
    type: object
  model.UserContent:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      phone:
        type: string
      projectRoles:
        items:
          $ref: '#/definitions/model.PROJECT_ROLE_TYPES'
        type: array
      scrumRoles:
        items:
          $ref: '#/definitions/model.SCRUM_ROLE_TYPES'
        type: array
    type: object
  model.UserContentResponse:
    properties:
      ID:
        type: string
      UUID:
        type: string
      content:
        $ref: '#/definitions/model.UserContent'
    type: object
  model.UserMetadata:
    properties:
      dates:
        $ref: '#/definitions/model.CommonDate'
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample API server using Gin and Swagger.
  title: My API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get Users
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      parameters:
      - description: New user details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create User
      tags:
      - User
  /users/{UUID}:
    delete:
      description: Delete User by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete User by UUID
      tags:
      - User
    get:
      description: Get User by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get User by UUID
      tags:
      - User
    put:
      description: Update User by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated user details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update User by UUID
      tags:
      - User
  /users/{UUID}/content:
    get:
      description: Get User Content by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User content found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get User Content by UUID
      tags:
      - User
    put:
      description: Update User Content by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated user content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserContentRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User content updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update User Content by UUID
      tags:
      - User
  /users/{UUID}/metadata:
    get:
      description: Get User Metadata by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User metadata found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get User Metadata by UUID
      tags:
      - User
    put:
      description: Update User Metadata by UUID
      parameters:
      - description: User UUID
        in: path
        name: UUID
        required: true
        type: string
      - description: Updated user metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserMetadataRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User metadata updated
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update User Metadata by UUID
      tags:
      - User
  /users/content:
    get:
      description: Get Users Content
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListUsersContentResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get Users Content
      tags:
      - User
  /users/email/{email}:
    get:
      description: Get User by Email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get User by Email
      tags:
      - User
  /users/id/{ID}:
    get:
      description: Get User by ID
      parameters:
      - description: User ID
        in: path
        name: ID
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - User
...
schemes:
- http
- https
swagger: "2.0"